requirements.txt
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dataclasses==0.7
fastapi==0.58.0
geographiclib==1.50
geopy==1.22.0
google-auth==1.17.2
google-auth-oauthlib==0.4.1
gspread==3.6.0
gunicorn==20.0.4
h11==0.9.0
httplib2==0.18.1
idna==2.9
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.5.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
starlette==0.13.4
tabulate==0.8.7
urllib3==1.25.9
uvicorn==0.11.5
websockets==8.1
